Move base presenter and view under base folder
