+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/MainContract.kt
@@ -1,5 +1,8 @@
 package com.bounswe2020group3.paperlayer
 
+import com.bounswe2020group3.paperlayer.base.BasePresenter
+import com.bounswe2020group3.paperlayer.base.BaseView
+
 interface MainContract {
   interface Presenter : BasePresenter {
     fun onViewCreated()
