+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/base/BasePresenter.kt
@@ -1,4 +1,4 @@
-package com.bounswe2020group3.paperlayer
+package com.bounswe2020group3.paperlayer.base
 
 interface BasePresenter {
   fun onDestroy()
