+++ android/app/src/main/java/com/bounswe2020group3/paperlayer/base/BaseView.kt
@@ -1,4 +1,4 @@
-package com.bounswe2020group3.paperlayer
+package com.bounswe2020group3.paperlayer.base
 
 interface BaseView<T> {
   fun setPresenter(presenter : T)
